capa field in RString, and RArray is now LONG (ruby-lang:458)
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@2828 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ ruby.h
@@ -316,7 +316,7 @@ struct RString {
     long len;
     char *ptr;
     union {
-	int capa;
+	long capa;
 	VALUE shared;
     } aux;
 };
@@ -325,7 +325,7 @@ struct RArray {
     struct RBasic basic;
     long len;
     union {
-	int capa;
+	long capa;
 	VALUE shared;
     } aux;
     VALUE *ptr;
